requirements.txt
asgiref==3.8.1
certifi==2024.2.2
charset-normalizer==3.3.2
Django==5.0.4
django-ckeditor==6.7.1
django-js-asset==2.2.0
djangorestframework==3.15.1
gunicorn==20.1.0
idna==3.6
pillow==10.3.0
psycopg2-binary==2.9.9
requests==2.31.0
sqlparse==0.4.4
urllib3==2.2.1
